+++ selfdrive/manager/manager.py
@@ -19,7 +19,7 @@ from openpilot.selfdrive.manager.helpers import unblock_stdout, write_onroad_par
 from openpilot.selfdrive.manager.process import ensure_running
 from openpilot.selfdrive.manager.process_config import managed_processes
 from openpilot.selfdrive.athena.registration import register, UNREGISTERED_DONGLE_ID
-from openpilot.system.swaglog import cloudlog, add_file_handler
+from openpilot.common.swaglog import cloudlog, add_file_handler
 from openpilot.system.version import is_dirty, get_commit, get_version, get_origin, get_short_branch, \
                            get_normalized_origin, terms_version, training_version, \
                            is_tested_branch, is_release_branch
@@ -37,6 +37,8 @@ def manager_init() -> None:
   params.clear_all(ParamKeyType.CLEAR_ON_MANAGER_START)
   params.clear_all(ParamKeyType.CLEAR_ON_ONROAD_TRANSITION)
   params.clear_all(ParamKeyType.CLEAR_ON_OFFROAD_TRANSITION)
+  if is_release_branch():
+    params.clear_all(ParamKeyType.DEVELOPMENT_ONLY)
 
   default_params: List[Tuple[str, Union[str, bytes]]] = [
     ("AleSato_AutomaticBrakeHold", "1"),
@@ -45,7 +47,7 @@ def manager_init() -> None:
     ("SshEnabled", "1"),
     ("GithubUsername", "AlexandreSato"),
     ("GithubSshKeys", "ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQCUfU4ymJdRned22jes0n0qI1vSSreusRsSS1pFcCroGGkfYU1ZTKuURGtBxcJSw5HEIdWRQfbmdGJBH/k+C1Y3hmgBkbJir+xwp28YeMdT8ZPEIzi6TQdJBee9LplHgxbrOP9M559Copf6lnyJdVQphKPOl91W/fOgC/xvKeS5v2CNiZCYgAIJsOtgIv1aw+wZVNhPTIda1Sv+6Gj5uk3YBvGcSwSvQBbXORJvaBuJAv0kVL0nLGv8OtQTmNrMsmR17+lHeVkkfFAaNE3E3QtPCCgpClp5FSbw5SpbYAqVk2MTIjfSGHDzYckT46l63gYHWSIHlqR5peIY/5hPUt6paDLdLwM50s5azwtMqZcZ+xY1QEu6wzQDTb+Z2JUm/VtjOmmnRBbXKuDNJ7HKOTvrmZcmbcr3wsTPC8VRHjrxR1TPoTwLYgwonaVlyFwP1W7KPYci6LG9xJraRQ/2W1zGupzmsoF3Zz5Uz01owrQCCpdRr2mbBSCupZjDgqs6lNhkWPPUEgkbm2t2nPCJyq0E0XjwmE+CrlPaE2SuKN3BJ3OmY46cyrtxmKgGfLxC+ZXFs10B4A2GCMuZhKm4WjlmfC8pDgr0PBMW/xYPLpYdEuDavPOXG/AViGHZlL0BIviihoZ1YbtuGFq2LTPL2IaM4l02x+Cqb2gztqc4mCVUGQ=="),
-    ("GsmMetered", "1"),
+    ("GsmMetered", "0"),
     ("HasAcceptedTerms", "0"),
     ("IsMetric", "1"),
     ("LanguageSetting", "main_pt-BR"),
@@ -174,7 +176,7 @@ def manager_thread() -> None:
     cloudlog.debug(running)
 
     # send managerState
-    msg = messaging.new_message('managerState')
+    msg = messaging.new_message('managerState', valid=True)
     msg.managerState.processes = [p.get_process_state_msg() for p in managed_processes.values()]
     pm.send('managerState', msg)