+++ RELEASES.md
@@ -1,7 +1,16 @@
-Version 0.9.6 (2023-XX-XX)
+Version 0.9.6 (20XX-XX-XX)
 ========================
 * AleSato stuff
- * Add QrCode to watch dashcam footage (2023-11-21)
+  * Add QrCode to watch dashcam footage (2023-11-21)
+* New driving model
+  * Vision model trained on more data
+  * Improved driving performance
+* AGNOS 9
+* comma body streaming and controls over WebRTC
+* Hyundai Staria 2023 support thanks to sunnyhaibin!
+* Kia Niro Plug-in Hybrid 2022 support thanks to sunnyhaibin!
+* Toyota RAV4 2023 support
+* Toyota RAV4 Hybrid 2023 support
 
 Version 0.9.5 (2023-11-17)
 ========================